Simplify era detection logic and address comments
